Publish role name in delete event.

+++ components/role-mgt/org.wso2.carbon.identity.role.v2.mgt.core/src/main/java/org/wso2/carbon/identity/role/v2/mgt/core/RoleManagementServiceImpl.java
@@ -399,8 +399,14 @@ public class RoleManagementServiceImpl implements RoleManagementService {
         RoleManagementEventPublisherProxy roleManagementEventPublisherProxy = RoleManagementEventPublisherProxy
                 .getInstance();
         roleManagementEventPublisherProxy.publishPreDeleteRoleWithException(roleId, tenantDomain);
+        String roleName = roleDAO.getRoleBasicInfoById(roleId, tenantDomain).getName();
+        RoleBasicInfo roleBasicInfo = roleDAO.getRoleBasicInfoById(roleId, tenantDomain);
+        if (roleBasicInfo == null) {
+            throw new IdentityRoleManagementException(String.format(
+                    "Role with id: %s does not exist in tenant domain: %s.", roleId, tenantDomain));
+        }
         roleDAO.deleteRole(roleId, tenantDomain);
-        roleManagementEventPublisherProxy.publishPostDeleteRole(roleId, tenantDomain);
+        roleManagementEventPublisherProxy.publishPostDeleteRole(roleId, roleName, tenantDomain);
         for (RoleManagementListener roleManagementListener : roleManagementListenerList) {
             if (roleManagementListener.isEnable()) {
                 roleManagementListener.postDeleteRole(roleId, tenantDomain);
@@ -948,10 +954,18 @@ public class RoleManagementServiceImpl implements RoleManagementService {
         }
         RoleManagementEventPublisherProxy roleManagementEventPublisherProxy = RoleManagementEventPublisherProxy
                 .getInstance();
-        List<String> deletedRoleIds = roleDAO.deleteRolesByApplicationAndReturnIds(applicationId, tenantDomain);
-        // Publish a post delete role event for each role that was deleted along with the application.
-        for (String deletedRoleId : deletedRoleIds) {
-            roleManagementEventPublisherProxy.publishPostDeleteRole(deletedRoleId, tenantDomain);
+        List<RoleBasicInfo> deletedRoles = roleDAO.deleteRolesByApplicationAndReturnRoles(applicationId,
+                tenantDomain);
+        // Publish a post delete role event (with the role name) for each role that was deleted along with the
+        // application.
+        for (RoleBasicInfo deletedRole : deletedRoles) {
+            roleManagementEventPublisherProxy.publishPostDeleteRole(deletedRole.getId(), deletedRole.getName(),
+                    tenantDomain);
+            for (RoleManagementListener roleManagementListener : roleManagementListenerList) {
+                if (roleManagementListener.isEnable()) {
+                    roleManagementListener.postDeleteRole(deletedRole.getId(), tenantDomain);
+                }
+            }
         }
         for (RoleManagementListener roleManagementListener : roleManagementListenerList) {
             if (roleManagementListener.isEnable()) {
